Complete refactor of code.
