+++ bin/isort-identify-imports
@@ -1,0 +1,8 @@
+#!/home/hina/Code/nerd-bot/bin/python
+# -*- coding: utf-8 -*-
+import re
+import sys
+from isort.main import identify_imports_main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(identify_imports_main())
