+++ bin/youtube-dl
@@ -1,0 +1,8 @@
+#!/home/hina/Code/nerd-bot/bin/python
+# -*- coding: utf-8 -*-
+import re
+import sys
+from youtube_dl import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
