+++ pyvenv.cfg
@@ -1,0 +1,8 @@
+home = /usr
+implementation = CPython
+version_info = 3.9.5.final.0
+virtualenv = 20.4.7
+include-system-site-packages = false
+base-prefix = /usr
+base-exec-prefix = /usr
+base-executable = /usr/bin/python
